feat: support variadic invokations of set
Support up to arity 7 invokations when using set as a function.

Note: I construct an anonymous arity 1 setter in the various impls of
invoke. The idea was to not make `-send` not need to be variadic itself.
This has the downside that `apply` can't be truly unbounded, but
simplifies the implementation of `-send` on source. It could be worth
considering making `-send` take an `args` vec (or make it variadic
itself).